requirements.txt
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cog==0.8.1
cogapp==3.3.0
datasets==2.14.0
diffusers==0.18.2
dill==0.3.7
fastapi==0.98.0
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
pyarrow==12.0.1
pydantic==1.10.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
structlog==23.1.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.2
